+++ src/papass/__init__.py
@@ -75,7 +75,7 @@ the ``alphabet`` like e.g ``'3002212230'``. The ``entropy`` is ``20.0`` because 
 from .passphrase_generator import PassphraseGenerator, PassphraseResult
 from .password_generator import PasswordGenerator, PasswordResult
 from .random import DiceRng, QueryForDice, RngBase, SystemRng
-from .utils import QueryStdinForDice
+from .utils import QueryUserForDice
 from .wordlist import WordList
 
 __version__ = "0.0.4"
@@ -87,7 +87,7 @@ __all__ = [
     PasswordGenerator.__name__,
     PasswordResult.__name__,
     QueryForDice.__name__,
-    QueryStdinForDice.__name__,
+    QueryUserForDice.__name__,
     RngBase.__name__,
     SystemRng.__name__,
     WordList.__name__,
